Refactor: move CallScheduler to golem.core.golem_async
